Overhaul compiler detection again.
All of the clever "try to find all of the compilers that we have"
logic is gone.  You can either list all the compilers you want to
run on the command line to driver.py, or you can accept the default,
which is any of "g++" and "clang++" that exist.  It still knows how
to detect necessary command line options to get C++11 mode and/or
libc++ turned on.

The big win here is that we no longer have to try to see through
symlinks and shell scripts.  The big lose is that we don't know how
to use g++ with libc++ anymore, but that was never all that reliable,
and it shouldn't be *that* hard to add back if I decide that I care
enough.

Also, once compilers have been probed, the set of working compilers +
switches is recorded in a file ("compilers.ini") in the source directory,
so you don't have to wait for it to re-probe them every single time you
run the test suite.

+++ test/util.py
@@ -1,58 +1,27 @@
 # Test utility functions.
 
 import contextlib
+import errno
 import os
-import subprocess
 import tempfile
 
 @contextlib.contextmanager
 def mkstemp_autodel(suffix="", prefix="tmp", dir=None, text=False,
                     contents=None):
-    (handle, pathname) = tempfile.mkstemp(suffix, prefix, dir, text)
-    if contents is not None:
-        os.write(handle, contents)
-    os.close(handle)
-    yield pathname
-    os.unlink(pathname)
-
-def is_script(path):
-    """Returns False if 'path' is not a script,
-       True if it's a script but not a shell script,
-       or the name of the shell interpreter if it's a shell script."""
-
-    with open(path, "rU") as f:
-        # check for #! before attempting to read up to a newline.
-        if f.read(2) != "#!": return False
-        interpreter = f.readline().split()[0]
-        if interpreter.endswith("sh") and not interpreter.endswith("csh"):
-            return interpreter
-        return True
-
-# This should be in subprocess, but it ain't.
-def check_io(*popenargs, **kwargs):
-    """Same as subprocess.check_output() but takes an additional argument,
-       'input', allowing you to pass a string to the subprocess's stdin.
-       You may not use the 'stdin' and 'input' arguments at the same time."""
-
-    if 'stdout' in kwargs:
-        raise ValueError('stdout argument not allowed, it will be overridden.')
-    if 'input' in kwargs:
-        if 'stdin' in kwargs:
-            raise ValueError("'stdin' and 'input' may not be used together.")
-        inputdata = kwargs['input']
-        del kwargs['input']
-        kwargs['stdin'] = subprocess.PIPE
-    else:
-        inputdata = None
-    process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
-    output, unused_err = process.communicate(inputdata)
-    retcode = process.poll()
-    if retcode:
-        cmd = kwargs.get("args")
-        if cmd is None:
-            cmd = popenargs[0]
-        raise subprocess.CalledProcessError(retcode, cmd, output=output)
-    return output
+    pathname = None
+    try:
+        (handle, pathname) = tempfile.mkstemp(suffix, prefix, dir, text)
+        if contents is not None:
+            os.write(handle, contents)
+        os.close(handle)
+        yield pathname
+    finally:
+        if pathname is not None:
+            try:
+                os.unlink(pathname)
+            except OSError, e:
+                if e.errno != errno.ENOENT:
+                    raise
 
 # credit to stackoverflow user 'Obtuse':
 # http://stackoverflow.com/a/6849299/388520
